Fix: clippy

+++ pumpkin/src/client/client_packet.rs
@@ -67,7 +67,7 @@ impl Client {
         name.len() <= 16
             && name
                 .chars()
-                .all(|c| c > 32 as u8 as char && c < 127 as u8 as char)
+                .all(|c| c > 32_u8 as char && c < 127_u8 as char)
     }
 
     pub fn handle_login_start(&mut self, server: &mut Server, login_start: SLoginStart) {
